rename api methods to fetcher
