+++ elementary/monitor/fetchers/sidebar/schema.py
